+++ projects/ng-wizi-bulma/src/lib/animated-card/animated-card-module.ts
@@ -1,0 +1,18 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {NwbCommonModule} from '../shared/common-module';
+import {NwbAnimatedCardComponent} from './animated-card.component';
+
+
+@NgModule({
+  imports: [
+    CommonModule,
+    NwbCommonModule,
+  ],
+  providers: [],
+  entryComponents: [],
+  declarations: [NwbAnimatedCardComponent],
+  exports: [NwbAnimatedCardComponent],
+})
+export class NwbAnimatedCardModule {
+}