+++ projects/ng-wizi-bulma/src/lib/animated-card/index.ts
@@ -1,0 +1,2 @@
+export * from './animated-card.component';
+export * from './animated-card-module';