+++ projects/ng-wizi-bulma/src/lib/dialog/dialog-module.ts
@@ -1,0 +1,19 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {NwbDialogComponent} from './dialog.component';
+import {NwbDialogService} from './dialog.service';
+import {NwbCommonModule} from '../shared/common-module';
+
+
+@NgModule({
+  imports: [
+    CommonModule,
+    NwbCommonModule,
+  ],
+  providers: [NwbDialogService],
+  entryComponents: [NwbDialogComponent],
+  declarations: [NwbDialogComponent],
+  exports: [NwbDialogComponent],
+})
+export class NwbDialogModule {
+}