+++ projects/ng-wizi-bulma/src/lib/dialog/index.ts
@@ -1,0 +1,3 @@
+export * from './dialog.component';
+export * from './dialog.service';
+export * from './dialog-module';