+++ projects/ng-wizi-bulma/src/lib/dropdown/dropdown.component.ts
@@ -1,6 +1,6 @@
 import {Component, ContentChildren, ElementRef, HostListener, Input, QueryList, ViewEncapsulation} from '@angular/core';
 import {ControlValueAccessor, NG_VALUE_ACCESSOR} from '@angular/forms';
-import {NwbOptionComponent} from '../option/option.component';
+import {NwbOptionComponent} from './option.component';
 import {Observable} from 'rxjs';
 
 @Component({