+++ projects/ng-wizi-bulma/src/lib/dropdown/index.ts
@@ -1,0 +1,3 @@
+export * from './dropdown.component';
+export * from './option.component';
+export * from './dropdown-module';