+++ projects/ng-wizi-bulma/src/lib/modal-search/index.ts
@@ -1,0 +1,3 @@
+export * from './modal-search.component';
+export * from './modal-search.service';
+export * from './modal-search-module';