+++ projects/ng-wizi-bulma/src/lib/shared/directives/index.ts
@@ -1,0 +1,2 @@
+export * from './debounce/debounce.directive';
+export * from './tooltip/tooltip.directive';