+++ projects/ng-wizi-bulma/src/lib/snackbar/index.ts
@@ -1,0 +1,3 @@
+export * from './snackbar.service';
+export * from './snackbar.component';
+export * from './snackbar-module';