+++ projects/ng-wizi-bulma/src/lib/tabs/tabs-module.ts
@@ -1,0 +1,25 @@
+import {CommonModule} from '@angular/common';
+import {NgModule} from '@angular/core';
+import {NwbTabComponent} from './tab.component';
+import {NwbTabsComponent} from './tabs.component';
+import {NwbCommonModule} from '../shared/common-module';
+
+
+@NgModule({
+  imports: [
+    CommonModule,
+    NwbCommonModule,
+  ],
+  providers: [],
+  entryComponents: [],
+  declarations: [
+    NwbTabsComponent,
+    NwbTabComponent
+  ],
+  exports: [
+    NwbTabsComponent,
+    NwbTabComponent
+  ],
+})
+export class NwbTabsModule {
+}