+++ src/app/app.module.ts
@@ -2,6 +2,8 @@ import {NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {FormsModule} from '@angular/forms';
+import {CdkTableModule} from '@angular/cdk/table';
+
 import {DialogDemo} from './features/dialog/dialog-demo';
 import {FakeDialogDemoComponent} from './features/dialog/fake-dialog-demo.component';
 import {SnackbarDemo} from './features/snackbar/snackbar-demo';
@@ -16,7 +18,7 @@ import {EditUserDialogDemoComponent} from './features/dialog/edit-user-dialog-de
 import {ModalSearchDemo} from './features/modal-search/modal-search-demo';
 import {AppComponent} from './app.component';
 import {AppRoutingModule} from './app-routing.module';
-import {NwbModule} from 'ng-wizi-bulma';
+
 import {SwitchDemo} from './features/switch/switch-demo';
 import {DocPreviewComponent} from './doc-preview/doc-preview.component';
 import {AlertDemo} from './features/alert/alert-demo';
@@ -25,7 +27,22 @@ import {AnimatedCardDemo} from './features/animated-card/animated-card-demo';
 import {TableDemo} from './features/table/table-demo';
 import {HttpClientModule} from '@angular/common/http';
 import {SortDemo} from './features/sort/sort-demo';
-import {CdkTableModule} from '@angular/cdk/table';
+
+import {
+  NwbAlertModule,
+  NwbAnimatedCardModule,
+  NwbCommonModule,
+  NwbDialogModule,
+  NwbDropdownModule,
+  NwbModalSearchModule,
+  NwbPaginatorModule,
+  NwbProgressBarModule,
+  NwbSnackbarModule,
+  NwbSortModule,
+  NwbSwitchModule,
+  NwbTabsModule
+} from 'ng-wizi-bulma';
+
 
 @NgModule({
   imports: [
@@ -33,9 +50,21 @@ import {CdkTableModule} from '@angular/cdk/table';
     BrowserAnimationsModule,
     FormsModule,
     HttpClientModule,
-    NwbModule,
     AppRoutingModule,
     CdkTableModule,
+
+    NwbCommonModule,
+    NwbAlertModule,
+    NwbAnimatedCardModule,
+    NwbDialogModule,
+    NwbDropdownModule,
+    NwbModalSearchModule,
+    NwbPaginatorModule,
+    NwbProgressBarModule,
+    NwbSnackbarModule,
+    NwbSortModule,
+    NwbSwitchModule,
+    NwbTabsModule,
   ],
   declarations: [
     AppComponent,